backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
httpx==0.27.0
pydantic==2.9.0